refactor hpc related components, remove hpc-admin stuff
