+++ garden/src/features/functions/hpc/utils/parseHpcFunctions.ts
