--- garden/src/features/hpc-admin/api/useHpcFunctions.ts
@@ -1,13 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import axios from "@/lib/axios";
-import { HpcFunctionMetadataResponse } from "@/types";
-
-export const useHpcFunctions = () => {
-  return useQuery({
-    queryKey: ["hpc-functions"],
-    queryFn: async () => {
-      const response = await axios.get("/hpc/functions");
-      return response.data as HpcFunctionMetadataResponse[];
-    },
-  });
-};
