--- garden/src/features/hpc-admin/components/index.ts
@@ -1,3 +1,0 @@
-export { default as HpcAdminDashboard } from "./HpcAdminDashboard";
-export { CreateHpcEndpointForm } from "./CreateHpcEndpointForm";
-export { CreateHpcFunctionForm } from "./CreateHpcFunctionForm";
